+++ src/fn/index.mjs
@@ -1,3 +1,5 @@
 export * from './get-font-file-data.mjs';
+export * from './get-unicode-range.mjs';
+export * from './make-os2-table.mjs';
 export * from './parse-os2-table.mjs';
 export * from './uncompress-table.mjs';
